+++ index.js
@@ -2,12 +2,11 @@
  * Vue Web Request Mediator Components.
  *
  * New BSD License (3-clause)
- * Copyright (c) 2018, Digital Bazaar, Inc.
+ * Copyright (c) 2018-2019, Digital Bazaar, Inc.
  * All rights reserved.
  */
 'use strict';
 
-export {requestStorageAccess} from './storageAccess.js';
 export {getWebAppManifestIcon} from './manifest.js';
 
 export function install(Vue, options) {